updates on comments, moving debounce and nested code to proper files
