+++ src/components/Facility/Investigations/ShowInvestigation.tsx
@@ -9,6 +9,7 @@ import * as Notification from "@/Utils/Notifications";
 import routes from "@/Utils/request/api";
 import request from "@/Utils/request/request";
 import useQuery from "@/Utils/request/useQuery";
+import { setNestedValueSafely } from "@/Utils/utils";
 
 const initialState = {
   changedFields: {},
@@ -91,31 +92,8 @@ export default function ShowInvestigation(props: ShowInvestigationProps) {
 
   const handleValueChange = (value: any, name: string) => {
     const changedFields = { ...state.changedFields };
-    const keys = name.split(".");
-    let current = changedFields;
-    for (let i = 0; i < keys.length - 1; i++) {
-      const key = keys[i];
 
-      // Protect against prototype pollution by skipping unsafe keys - crai
-      if (key === "__proto__" || key === "constructor" || key === "prototype") {
-        continue;
-      }
-
-      // Use Object.create(null) to prevent accidental inheritance from Object prototype - coderabbit
-      current[key] = current[key] || Object.create(null);
-      current = current[key];
-    }
-
-    const lastKey = keys[keys.length - 1];
-
-    // Final key assignment, ensuring no prototype pollution vulnerability - coderabbit
-    if (
-      lastKey !== "__proto__" &&
-      lastKey !== "constructor" &&
-      lastKey !== "prototype"
-    ) {
-      current[lastKey] = value;
-    }
+    setNestedValueSafely(changedFields, name, value);
 
     dispatch({ type: "set_changed_fields", changedFields });
   };
